Fixed issue where trace was read after the device was destroyed cleanly by the user

+++ src/runtime_src/xdp/profile/plugin/device_offload/device_offload_plugin.h
@@ -60,7 +60,7 @@ namespace xdp {
                        TraceLoggerCreatingDeviceEvents*,
                        DeviceIntf*> DeviceData ;
 
-    std::map<uint32_t, DeviceData> offloaders ;
+    std::map<uint64_t, DeviceData> offloaders ;
 
     XDP_EXPORT void addDevice(const std::string& sysfsPath) ;
     XDP_EXPORT void configureDataflow(uint64_t deviceId, DeviceIntf* devInterface) ;
